requirements.txt
python-telegram-bot==20.3
python-docx
python-slugify
requests
newspaper3k
lxml[html_clean]
